Update neighbor exploration in extension
A net that is used to connect one neighbor cannot be used to connect
another one, unless multiple pins of the MOS connect to it.
After the update, the result path seems to still be BROKEN, as it's hard
to identify which nets are free with the current data structure.